docs: 修改注释

+++ raft-kvstore/src/main/kotlin/com/qianlei/kv/server/config/ServerConfig.kt
@@ -14,11 +14,11 @@ data class ServerConfig(
     val selfId: NodeId,
     val dataPath: String = "data",
     /* 最小选举超时时间 */
-    val minElectionTimeout: Int = 2000,
+    val minElectionTimeout: Long = 2000,
     /* 最大选举超时时间 */
-    val maxElectionTimeout: Int = 4000,
+    val maxElectionTimeout: Long = 4000,
     /* 初次日志复制延迟时间 */
-    val logReplicationDelay: Int = 1000,
+    val logReplicationDelay: Long = 1000,
     /* 日志复制间隔 */
-    val logReplicationInterval: Int = 200,
+    val logReplicationInterval: Long = 200,
 )
